feat(inventory): true fractional stock (sell/receive by weight)
Migrate stock from integer units to real values so a kg-tracked item can be
stocked, received and sold in fractions (1.5 kg) with stock decrementing 1.5.

- Schema v14: inventory.stock + stock_movements.delta/balance_after -> REAL
  (SQLite numeric affinity keeps existing whole values; transparent upgrade)
- Model types flipped int->double across the chain: InventoryCatalogItem,
  PosCartItem, LowStockItem, SaleDetailItem, StockMovement, ReportSaleLine,
  TopProduct, VariantDraft, LocalSaleCommit.inventoryDeltas
- Repo: stock reads/writes, applyStockIn, logStockAdjustment, sale decrement,
  return restock and movement deltas are all fractional
- POS: qty steppers/cart carry double; tap a cart line's quantity to type an
  exact fractional amount (e.g. 1.5); opening stock, restock and purchase
  received-qty accept decimals
- Money math stays exact: price*qty rounds to paise via Money; quantities are
  shown without trailing ".0" (formatQty)
- COGS/GST/P&L, low-stock threshold, Z-report all work on fractional qty

NOTE: the Django sale serializer must accept fractional quantity before cloud
sync is enabled for weighed items (backend sync currently local-first/off).

test-first: fractional_stock_test.dart (line totals, mixed carts, COGS,
low-stock); existing 137 tests updated + still green (141 total).
Verified: analyze clean (new code), 141 tests pass, release APK 83.9MB.

+++ apps/mobile_flutter/lib/core/backend/backend_api_client.dart
@@ -336,7 +336,7 @@ class BackendApiClient {
     required String shopId,
     required String name,
     required double sellPrice,
-    required int openingStock,
+    required double openingStock,
     String sku = '',
     String barcode = '',
     String category = 'General',
